List all polling stations

Retrieve list of all polling station that been used from 2044 onwards. Names as recorded by the AEC. List can be filtered by state, division names and regular expressions matching their names.

Usage

list_polling_places(filters = NULL)

Arguments

filters

(optional) list() with filters in the form list(Column="Value")

Value

data frame with lists of polling stations

Examples

# \dontrun{
# Get list of all registered parties
list_parties()
#> # A tibble: 387 × 10
#>    StateAb PartyAb PartyNm      `2016` `2019` `2004` `2010` `2013` `2022` `2007`
#>    <chr>   <chr>   <chr>        <lgl>  <lgl>  <lgl>  <lgl>  <lgl>  <lgl>  <lgl> 
#>  1 NSW     AAPP    Antipaedoph… TRUE   NA     NA     NA     NA     NA     NA    
#>  2 NSW     ABFA    Australian … NA     TRUE   NA     NA     NA     NA     NA    
#>  3 VIC     ADP     The Aged an… NA     NA     TRUE   NA     NA     NA     NA    
#>  4 NSW     ADVP    Veterans Pa… TRUE   NA     NA     NA     NA     NA     NA    
#>  5 QLD     ADVP    Veterans Pa… TRUE   NA     NA     NA     NA     NA     NA    
#>  6 VIC     AEQ     Marriage Eq… TRUE   NA     NA     NA     NA     NA     NA    
#>  7 NSW     AFN     Australia F… TRUE   TRUE   NA     TRUE   TRUE   NA     NA    
#>  8 NT      AFN     Australia F… TRUE   NA     NA     NA     NA     NA     NA    
#>  9 QLD     AFN     Australia F… NA     TRUE   NA     NA     NA     NA     NA    
#> 10 SA      AFN     Australia F… NA     NA     NA     NA     TRUE   NA     NA    
#> # … with 377 more rows
# Get list of polling places in the division of Hasluck
list_parties(list)
#> # A tibble: 387 × 10
#>    StateAb PartyAb PartyNm      `2016` `2019` `2004` `2010` `2013` `2022` `2007`
#>    <chr>   <chr>   <chr>        <lgl>  <lgl>  <lgl>  <lgl>  <lgl>  <lgl>  <lgl> 
#>  1 NSW     AAPP    Antipaedoph… TRUE   NA     NA     NA     NA     NA     NA    
#>  2 NSW     ABFA    Australian … NA     TRUE   NA     NA     NA     NA     NA    
#>  3 VIC     ADP     The Aged an… NA     NA     TRUE   NA     NA     NA     NA    
#>  4 NSW     ADVP    Veterans Pa… TRUE   NA     NA     NA     NA     NA     NA    
#>  5 QLD     ADVP    Veterans Pa… TRUE   NA     NA     NA     NA     NA     NA    
#>  6 VIC     AEQ     Marriage Eq… TRUE   NA     NA     NA     NA     NA     NA    
#>  7 NSW     AFN     Australia F… TRUE   TRUE   NA     TRUE   TRUE   NA     NA    
#>  8 NT      AFN     Australia F… TRUE   NA     NA     NA     NA     NA     NA    
#>  9 QLD     AFN     Australia F… NA     TRUE   NA     NA     NA     NA     NA    
#> 10 SA      AFN     Australia F… NA     NA     NA     NA     TRUE   NA     NA    
#> # … with 377 more rows

 # }